Remove old migrater type
